refactor fix and added more route according to the db schema changes in workexperience

+++ src/modules/work/handlers/delete-position-experience-by-id.handler.ts
@@ -6,17 +6,18 @@ import { customZValidator } from "@/middlewares/custom-z-validator";
 import { HTTPException } from "hono/http-exception";
 import z from "zod";
 
-export const deleteExperienceById = factory.createHandlers(
+export const deletePositionExperienceById = factory.createHandlers(
   customZValidator(
     "param",
     z.object({
       id: z.string(),
+      positionId: z.string(),
     }),
   ),
 
   async (c) => {
     try {
-      const { id } = c.req.valid("param");
+      const { id, positionId } = c.req.valid("param");
 
       const experience = await WorkExperienceModel.findById(id);
       if (!experience) {
@@ -25,25 +26,40 @@ export const deleteExperienceById = factory.createHandlers(
         });
       }
 
-      await WorkExperienceModel.findByIdAndDelete(id);
+      const position = experience.positions.id(positionId);
+      if (!position) {
+        throw new HTTPException(StatusCodes.HTTP_400_BAD_REQUEST, {
+          message: `Position with id ${positionId} does not exist`,
+        });
+      }
+
+      position.deleteOne();
+
+      let result;
+      if (experience.positions.length === 0) {
+        result = await WorkExperienceModel.findByIdAndDelete(id);
+      } else {
+        result = await experience.save();
+      }
 
       return c.json(
         {
-          message: "work expirence deleted successfully",
+          message: "Position deleted successfully",
+          data: result,
         },
         StatusCodes.HTTP_200_OK,
       );
     } catch (err) {
       if (err instanceof HTTPException) throw err;
 
-      logger.error("Error while deleting work experience", {
+      logger.error("Error while deleting work position experience", {
         module: "work",
         action: "work:delete:error",
         error: err instanceof Error ? err.message : String(err),
       });
 
       throw new HTTPException(StatusCodes.HTTP_500_INTERNAL_SERVER_ERROR, {
-        message: "Failed to delete work exp",
+        message: "Failed to delete work position exp",
       });
     }
   },
